Implemento las nuevas funciones de Protocol

+++ protocol/MapInfo.h
@@ -4,13 +4,13 @@
 #include <list>
 
 struct Box{
-    int x;
-    int y;
+    uint16_t x;
+    uint16_t y;
 };
 
 struct MapInfo{
-    int height;
-    int length;
+    uint16_t height;
+    uint16_t length;
     std::list<Box> boxes;
 };
